% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-impala.R
\name{db_desc}
\alias{db_desc}
\alias{db_desc.impala_connection}
\title{Describe the Impala data source}
\usage{
\method{db_desc}{impala_connection}(x)
}
\arguments{
\item{x}{an object with class class \code{impala_connection}}
}
\value{
A string containing information about the connection to Impala
}
\description{
Describe the Impala data source
}
