% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_impala.R
\name{src_desc}
\alias{src_desc}
\alias{src_desc.src_impala}
\title{Describe the Impala data source}
\usage{
\method{src_desc}{src_impala}(x)
}
\arguments{
\item{x}{an object with class class \code{src_impala}}
}
\value{
A string containing information about the connection to Impala
}
\description{
Describe the Impala data source
}
