% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_impala.R
\name{tbl}
\alias{tbl}
\alias{tbl.src_impala}
\title{Create a lazy \code{tbl} from an Impala table}
\usage{
\method{tbl}{src_impala}(src, from, ...)
}
\arguments{
\item{src}{an object with class with class \code{src_impala}}

\item{from}{a table name}

\item{...}{not used}
}
\value{
An object with class \code{tbl_impala}, \code{tbl_sql},
  \code{tbl_lazy}, \code{tbl}
}
\description{
Create a lazy \code{tbl} from an Impala table
}
\examples{
\dontrun{
flights_tbl <- tbl(impala, "flights")}
}
